Content Helper: Move "Current Post Details" files to own folder
