+++ src/blocks/content-helper/current-post-details/data.ts
